requirements.txt
alembic==1.13.1
attrs==23.2.0
bcrypt==4.1.2
blinker==1.7.0
cachelib==0.9.0
cffi==1.16.0
click==8.1.7
cryptography==42.0.5
flasgger==0.9.7.1
Flask==3.0.2
Flask-Bcrypt==1.0.1
Flask-Caching==2.1.0
Flask-Cors==4.0.0
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
gunicorn==21.2.0
importlib_metadata==7.1.0
itsdangerous==2.1.2
Jinja2==3.1.3
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
Mako==1.3.2
MarkupSafe==2.1.5
mistune==3.0.2
packaging==24.0
psycopg2-binary==2.9.9
pycparser==2.22
PyMySQL==1.1.0
python-dotenv==1.0.1
PyYAML==6.0.1
referencing==0.34.0
rpds-py==0.18.0
six==1.16.0
SQLAlchemy==2.0.29
typing_extensions==4.10.0
Werkzeug==3.0.1
zipp==3.18.1
